backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.1.1
openpyxl==3.1.5
google-auth==2.25.2
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
requests==2.32.3
dj-database-url==2.2.0
